% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{auc.b}
\alias{auc.b}
\title{auc.b}
\usage{
auc.b(b.est, b.true, is.association = FALSE, ...)
}
\arguments{
\item{b.est}{The estimated interaction network (This can be a correlation matrix and the edges will be ranked based on the absolute value of the correlation coefficents)}

\item{b.true}{The true interaction network}

\item{is.association}{b.est is an association structure}
}
\description{
plot ROC curve with AUC for the interaction graph (interaction signs are ignored)
}
\author{
Chenhao Li, Niranjan Nagarajan
}
