% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beem2param.r
\name{beem2biomass}
\alias{beem2biomass}
\title{beem2biomass}
\usage{
beem2biomass(beem)
}
\arguments{
\item{beem}{a BEEM object}
}
\description{
extract biomass estimates from a BEEM object
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
