% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beem2param.r
\name{beem2param}
\alias{beem2param}
\title{beem2param}
\usage{
beem2param(beem)
}
\arguments{
\item{beem}{a BEEM object}
}
\description{
extract parameter estimates from a BEEM object
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
