% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beemDemo.r
\docType{data}
\name{beemDemo}
\alias{beemDemo}
\title{Demo dataset for BEEM}
\format{A list containing the demo data for BEEM-{Static}
\describe{
  \item{dat.w.noise}{Counts data with Poisson noise}
  \item{dat.wo.noise}{Raw data simulated from a GLVM}
  \item{scaled.params}{parameters scaled (by dividing the self-interaction term)}
  \item{biomass.true}{true biomass values}
  ...
}}
\usage{
beemDemo
}
\description{
Demo dataset for BEEM
}
\keyword{datasets}
