% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizations.r
\name{biomassInit}
\alias{biomassInit}
\title{biomassInit}
\usage{
biomassInit(m)
}
\arguments{
\item{m}{matrix of data (variables in columns, measurements in rows)}
}
\description{
Function to initialize biomass. This normalization assumes that all samples are at the equilibrium and there is no interaction at all.
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
