% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{checkPackage}
\alias{checkPackage}
\title{checkPackage}
\usage{
checkPackage(x)
}
\arguments{
\item{x}{package to check}
}
\description{
Check if a package is available
}
\author{
Chenhao Li, Niranjan Nagarajan
}
