% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{cluster}
\alias{cluster}
\title{cluster}
\usage{
cluster(countData)
}
\arguments{
\item{countData}{OTU/species abundance table (each row is one species, each column is one site)}
}
\description{
perform a hierachical clustering analysis using Bray-Curtis distance
}
\author{
Chenhao Li, Niranjan Nagarajan
}
