% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{css}
\alias{css}
\title{css}
\usage{
css(m, p = 0.5)
}
\arguments{
\item{m}{matrix of data (variables in columns, measurements in rows)}

\item{p}{quantile used for normalization (default: 0.5)}
}
\description{
Function to perform cumulative sum scaling (CSS)
}
\author{
Chenhao Li, Niranjan Nagarajan
}
