% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{detectBadSamples}
\alias{detectBadSamples}
\title{detectBadSamples}
\usage{
detectBadSamples(err, threshold)
}
\arguments{
\item{err}{the errors estimated from regression}

\item{threshold}{threshold to filter out samples}
}
\description{
detectBadSamples
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
