% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.EM.r
\name{detectDiffModSamples}
\alias{detectDiffModSamples}
\title{detectDiffModSamples}
\usage{
detectDiffModSamples(err, threshold, filter)
}
\arguments{
\item{err}{the errors defined in `func.EM`}

\item{threshold}{threshold to filter out samples}
}
\description{
detectDiffModSamples
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
