% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.EM.r
\name{detectNonEqSamples}
\alias{detectNonEqSamples}
\title{detectNonEqSamples}
\usage{
detectNonEqSamples(err, threshold)
}
\arguments{
\item{err}{the errors defined in `func.EM`}

\item{threshold}{threshold to filter out samples}
}
\description{
detectNonEqSamples
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
