% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{diagnoseBiomass}
\alias{diagnoseBiomass}
\title{diagnoseBiomass}
\usage{
diagnoseBiomass(beem.out, true.biomass = NA, alpha = 0.1, ...)
}
\arguments{
\item{beem.out}{output of a beem run}

\item{true.biomass}{measured/true biomass (default: the biomass of last iteration)}

\item{alpha}{transparency parameter of the biomass lines}
}
\description{
plot the trace of the fitted error of biomass
}
\author{
Chenhao Li, Niranjan Nagarajan
}
