% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{diagnoseFit}
\alias{diagnoseFit}
\title{diagnoseFit}
\usage{
diagnoseFit(beem.out, dat, thre = 0.5, annotate = TRUE)
}
\arguments{
\item{beem.out}{output of a beem run}

\item{dat}{input data for the beem run}

\item{thre}{threshold of R2 to show the label}

\item{annotate}{display the label for the species with R2 > thre}
}
\description{
plot the R2 value for each species
}
\author{
Chenhao Li, Niranjan Nagarajan
}
