% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{entropy}
\alias{entropy}
\title{entropy}
\usage{
entropy(v)
}
\arguments{
\item{v}{a vector input (treated as binary, i.e. zero/non-zero)}
}
\description{
Compute the entropy of a binary vector
}
\author{
Chenhao Li, Niranjan Nagarajan
}
