% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{formatOutput}
\alias{formatOutput}
\title{formatOutput}
\usage{
formatOutput(a, b, vnames)
}
\arguments{
\item{a}{scaled growth rates}

\item{b}{scaled interaction matrix}

\item{vnames}{variable names}
}
\description{
Function to convert parameter vector a and matrix b to MDSINE's output format
}
