% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{func.E.stab}
\alias{func.E.stab}
\title{func.E.stab}
\usage{
func.E.stab(dat.tss, m, sample.filter, ncpu = 4, perc = 0.6,
  niter = 100, ...)
}
\arguments{
\item{dat.tss}{relative abundances matrix (each OTU in one row)}

\item{m}{estimated biomass values}

\item{sample.filter}{filter out samples contain outliers in Y}

\item{ncpu}{number of CPUs (default: 4)}

\item{perc}{percentage of samples to take for each iteration}

\item{niter}{number of iterations to run}

\item{...}{additional parameters for beemStatic:::infer}
}
\description{
E-step with subsampling to estimate stability
}
\author{
Chenhao Li, Niranjan Nagarajan
}
