% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{func.EM}
\alias{func.EM}
\title{func.EM}
\usage{
func.EM(dat, ncpu = 4, scaling = 10000, dev = 2, max.iter = 30,
  warm.iter = NULL, lambda.choice = 1)
}
\arguments{
\item{dat}{OTU count/relative abundance matrix (each OTU in one row)}

\item{ncpu}{number of CPUs (default: 4)}

\item{lambda.choice}{1: use lambda.1se for LASSO, 2: use lambda.min for LASSO}
}
\description{
Iteratively estimating scaled parameters and biomass
}
