% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{infer}
\alias{infer}
\title{infer}
\usage{
infer(Y, X, method = "glmnet", intercept = FALSE, seed = 0,
  alpha = 1, nfolds = 3, lambda.init = NULL, lambda.choice = 1,
  lambda.adjust.up = TRUE, lambda.adjust.down = TRUE)
}
\arguments{
\item{Y}{response}

\item{X}{predictors}

\item{method}{blasso or lm}

\item{intercept}{whether to include the intercept}

\item{seed}{seed}

\item{alpha}{the alpha parameter for elastic net (1:lasso [default], 0:ridge)}

\item{nfolds}{number of folds for glmnet cross-validation}

\item{lambda.init}{user provides initial lambda values}

\item{lambda.choice}{1: use lambda.1se for LASSO, 2: use lambda.min for LASSO, a number between (0, 1): this will select a lambda according to (1-lambda.choice)*lambda.min + lambda.choice*lambda.1se}

\item{lambda.adjust.up}{allow adjusting lambda (increase) by 50\% in each iteration}

\item{lambda.adjust.down}{allow adjusting lambda (decrease) by 50\% in each iteration}
}
\description{
Infer parameters with blasso
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
