% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.r
\name{inference}
\alias{inference}
\title{inference}
\usage{
inference(dat, beem, ncpu = 1, p.values = FALSE)
}
\arguments{
\item{dat}{original data used for BEEM-static prediction}

\item{beem}{BEEM-static ouput}

\item{ncpu}{number of CPUs (default: 1)}

\item{p.values}{report p-values computed from the confidence scores (default: FALSE)}
}
\description{
Infer confidence based on stepwise variable selection
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
