% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{interaction2association}
\alias{interaction2association}
\title{interaction2association}
\usage{
interaction2association(m)
}
\arguments{
\item{m}{an interaction matrix}
}
\description{
Convert interaction matrix to association matrix by taking the average of the symmetric entries
}
\author{
Chenhao Li, Niranjan Nagarajan
}
