% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{pcoa}
\alias{pcoa}
\title{pcoa}
\usage{
pcoa(countData, col = "Color", k = 2)
}
\arguments{
\item{countData}{OTU/species abundance table (each row is one species, each column is one site)}

\item{col}{a vector of colors for the points}

\item{k}{the number of dimensions}
}
\description{
perform a PCoA analysis using Bray-Curtis distance
}
\author{
Chenhao Li, Niranjan Nagarajan
}
