% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{postProcess}
\alias{postProcess}
\title{postProcess}
\usage{
postProcess(b, dev = 1e-05)
}
\arguments{
\item{b}{scaled interaction matrix (diagonal/self-interaction is -1)}

\item{dev}{scale of the interaction < dev * 1, the value will be set to 0}
}
\description{
post-process parameters to remove small entries
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
