% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.E.r
\name{postProcess_perturbation}
\alias{postProcess_perturbation}
\title{postProcess_perturbation}
\usage{
postProcess_perturbation(c, growth.rate, dev = 1e-05)
}
\arguments{
\item{dev}{scale of the interaction < dev * 1, the value will be set to 0}

\item{b}{scaled perturbation matrix (comparing to growthrate)}
}
\description{
post-process parameters to remove small entries
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
