% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.r
\name{predict.beem}
\alias{predict.beem}
\title{predict.beem}
\usage{
\method{predict}{beem}(object, dat.new, dev, ncpu = 4, pert.new = NULL)
}
\arguments{
\item{object}{output of the EM algorithm}

\item{dat.new}{OTU count/relative abundance matrix (each OTU in one row)}

\item{dev}{deviation of the error (for one sample) from the model to be excluded}

\item{ncpu}{number of CPUs (default: 4)}

\item{pert.new}{external perturbation presence matrix (each perturbation in one row, each sample in one column) (Default: NULL)}
}
\description{
Use a trained BEEM-static model to predict biomass, deviation from steady states and violation of model assumption
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
