% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.r
\name{predict_eq}
\alias{predict_eq}
\title{predict_eq}
\usage{
predict_eq(dat, a, b, m, ncpu = 1)
}
\arguments{
\item{dat}{OTU count/relative abundance matrix (each OTU in one row)}

\item{a}{a values}

\item{b}{b values}

\item{m}{biomass}

\item{ncpu}{number of CPUs (default: 1)}
}
\description{
predict the expected equilibria for a dataset
}
\author{
Chenhao Li
}
