% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{preProcess}
\alias{preProcess}
\title{preProcess}
\usage{
preProcess(dat, dev = 0)
}
\arguments{
\item{dat}{OTU count/relative abundance matrix (each OTU in one row)}

\item{dev}{dev * IQR from median will be filtered out (default: 0, nothing to remove)}
}
\description{
pre-process data
}
\author{
Chenhao Li, Niranjan Nagarajan
}
