% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.r
\name{print.beem}
\alias{print.beem}
\title{print.beem}
\usage{
\method{print}{beem}(object)
}
\arguments{
\item{object}{output of the EM algorithm}
}
\description{
print method for beem object
}
\author{
Chenhao Li
}
