% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{showGrowth}
\alias{showGrowth}
\title{showGrowth}
\usage{
showGrowth(beem.out, dat)
}
\arguments{
\item{beem.out}{output of a beem run}

\item{dat}{input data for the beem run}
}
\description{
ordered barplot for growth rates
}
\author{
Chenhao Li, Niranjan Nagarajan
}
