% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{sub_stat}
\alias{sub_stat}
\title{sub_stat}
\usage{
sub_stat(data, perc, ...)
}
\arguments{
\item{data}{data in the format of cbind(Y, X)}

\item{perc}{percentage of samples to take for each iteration}

\item{...}{additional parameters for beemStatic:::infer}
}
\description{
Resampling for the inference process
}
\author{
Chenhao Li, Niranjan Nagarajan
}
