% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{auc.b}
\alias{auc.b}
\title{auc.b}
\usage{
auc.b(b.est, b.true, is.association = FALSE, show = TRUE, ...)
}
\arguments{
\item{b.est}{the estimated interaction network (This can be a correlation matrix and the edges will be ranked based on the absolute value of the correlation coefficents)}

\item{b.true}{the true interaction network}

\item{is.association}{`b.est` is an association structure}

\item{show}{plot ROC curve. If set to `FALSE`, this will return the AUC value.}

\item{...}{additional arguments for `pROC::plot.roc`}
}
\description{
plot ROC curve with AUC for the interaction graph (interaction signs are ignored)
}
\author{
Chenhao Li, Niranjan Nagarajan
}
