% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{func.E}
\alias{func.E}
\title{func.E}
\usage{
func.E(dat.tss, external.perturbation = NULL, m, sample.filter,
  lambda.inits = NULL, ncpu = 4, center = FALSE, ...)
}
\arguments{
\item{dat.tss}{relative abundances matrix (each OTU in one row)}

\item{external.perturbation}{external perturbation presence matrix *1/m (each perturbation in one row, each sample in one column) (Default: NULL)}

\item{m}{estimated biomass values (1 X no. of samples) matrix}

\item{sample.filter}{filter out samples contain outliers in Y}

\item{lambda.inits}{initial lambda values}

\item{ncpu}{number of CPUs (default: 4)}

\item{center}{center data or not}

\item{...}{additional parameters for `beemStatic::infer`}
}
\description{
E-part of BEEM, estimate model parameters with inferred m
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
