% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.EM.r
\name{func.EM}
\alias{func.EM}
\title{func.EM}
\usage{
func.EM(dat, external.perturbation = NULL, ncpu = 1, m.init = NULL,
  scaling = 1000, equil.filter = Inf, model.filter = Inf,
  refresh.iter = 1, lambda.iter = Inf, warm.iter = NULL,
  max.iter = 30, epsilon = 0.001, lambda.choice = 1, alpha = 1,
  debug = FALSE, verbose = TRUE)
}
\arguments{
\item{dat}{OTU count/relative abundance matrix (each OTU in one row)}

\item{external.perturbation}{external perturbation presence matrix (each perturbation in one row, each sample in one column) (Default: NULL)}

\item{ncpu}{number of CPUs (default: 1)}

\item{m.init}{initial biomass values (default: use CSS normalization)}

\item{scaling}{a scaling factor to keep the median of all biomass constant (default: 1000)}

\item{equil.filter}{threshold for detecting and removing samples not at equilibrium (default: Inf - all the samples will be considered)}

\item{model.filter}{threshold for detecting and removing samples from different models (default: Inf - all the samples will be considered)}

\item{refresh.iter}{refresh the removed samples every X iterations (default: 1)}

\item{lambda.iter}{number of iterations to run before fixing lambda (default: Inf)}

\item{warm.iter}{number of iterations to run before removing any samples (default: run until convergence and start to remove samples)}

\item{max.iter}{maximal number of iterations (default 30)}

\item{epsilon}{convergence threshold (in relative difference): uqn of the relative error in biomass changes (default 1e-3)}

\item{lambda.choice}{1: use lambda.1se for LASSO, 2: use lambda.min for LASSO, a number between (0, 1): this will select a lambda according to (1-lambda.choice)*lambda.min + lambda.choice*lambda.1se}

\item{alpha}{the alpha parameter for the Elastic Net model (1-LASSO [default], 0-RIDGE)}

\item{debug}{output debugging information (default FALSE)}

\item{verbose}{print out messages}
}
\description{
Iteratively estimating scaled parameters and biomass
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
