% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{func.M}
\alias{func.M}
\title{func.M}
\usage{
func.M(dat.tss, a, b, ncpu = 4, ...)
}
\arguments{
\item{dat.tss}{relative abundance matrix (each OTU in one row)}

\item{a}{estimated growth rate values (scaled by self-interaction)}

\item{b}{estimated interaction matrix (scaled by self-interaction)}

\item{ncpu}{number of CPUs (default: 4)}
}
\description{
M-part of BEEM, estimate biomass with inferred model parameters
}
