% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{infer}
\alias{infer}
\title{infer}
\usage{
infer(Y, X, method = "glmnet", intercept = FALSE, seed = 0,
  lambda.choice = 1)
}
\arguments{
\item{Y}{response}

\item{X}{predictors}

\item{method}{blasso or lm}

\item{intercept}{whether to include the intercept}

\item{seed}{seed}

\item{lambda.choice}{1: use lambda.1se for analysis, 2: use lambda.min for analysis}
}
\description{
Infer parameters with blasso
}
