% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{norm}
\alias{norm}
\title{norm}
\usage{
norm(a, b, p = NULL, s = NULL, x)
}
\arguments{
\item{a}{estimated growth rate values (scaled by self-interaction)}

\item{b}{estimated interaction matrix (scaled by self-interaction)}

\item{p}{estimated external perturbation effect matrix (scaled by self-interaction) (default: NULL)}

\item{s}{external perturbation presence vector in one sample}

\item{x}{relative abundances in one sample}
}
\description{
estimate biomass with linear regression
}
\author{
Chenhao Li, Niranjan Nagarajan
}
