% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.r
\name{norm}
\alias{norm}
\title{norm}
\usage{
norm(a, b, x)
}
\arguments{
\item{a}{estimated growth rate values (scaled by self-interaction)}

\item{b}{estimated interaction matrix (scaled by self-interaction)}

\item{x}{abundances in one sample}
}
\description{
estimate biomass with linear regression
}
\author{
Chenhao Li, Niranjan Nagarajan
}
