% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{resample.EM}
\alias{resample.EM}
\title{resample.EM}
\usage{
resample.EM(data, external.perturbation = NULL, m, perc, res.iter, ...)
}
\arguments{
\item{data}{data in the format of cbind(Y, X)}

\item{external.perturbation}{external.perturbation presence matrix (each perturbation in one row, each sample in one column) (Default: NULL)}

\item{m}{biomass initialization}

\item{perc}{percentage of samples to take for each iteration}

\item{res.iter}{number of resample iteration}

\item{...}{additional parameters for beemStatic::func.EM}
}
\description{
Resampling for the inference process
}
\author{
Chenhao Li, Niranjan Nagarajan
}
