% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{showInteraction}
\alias{showInteraction}
\title{showInteraction}
\usage{
showInteraction(beem.out, dat, layout = "fr", node.text.size = 2,
  t.strength = 0.001, t.stab = 0.8, node.text.repel = TRUE)
}
\arguments{
\item{beem.out}{output of a beem run}

\item{dat}{input data for the beem run}

\item{layout}{graph layout (see the documentation for ggraph)}

\item{node.text.size}{node text size}

\item{t.strength}{the threshold used to limit the number of interactions (strength)}

\item{t.stab}{the threshold used to limit the number of interactions (stability)}

\item{node.text.repel}{use ggrepel to arrange node labels}
}
\description{
plot the interaction network inferred by beem (using ggraph)
}
\author{
Chenhao Li, Niranjan Nagarajan
}
