% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizations.r
\name{tss}
\alias{tss}
\title{tss}
\usage{
tss(x)
}
\arguments{
\item{x}{count matrix (each OTU in one row)}
}
\description{
Calculate the relative abundances by dividing the total abundance (Total Sum Sacling)
}
\author{
Chenhao Li, Gerald Tan, Niranjan Nagarajan
}
